Added the new files

+++ remove.php
@@ -1,11 +1,15 @@
 <?php
 require_once("services/database.php");
 
+if(!isset($_GET['email'])) {
+    die("There is an error with your request.");
+}
+
 if (AuthenticationService::delete($_GET['email'])) {
     echo "Success! Going back to homepage...";
     echo <<<END
         <script>
-            setTimeout(() => document.location.replace('/'), 3000);
+            setTimeout(() => document.location.replace('/'), 1000);
         </script>
     END;
 } else {
